cleancode: extract public func
